feitos métodos para validar diferentes apdex

+++ src/main/java/org/example/Apdex.java
@@ -2,7 +2,7 @@ package org.example;
 
 public class Apdex {
 
-    public float calcularApdex(float quantSatisfatorio, float quantToleravel, int totalAmostras){
+    public double calcularApdex(float quantSatisfatorio, float quantToleravel, int totalAmostras){
         return (quantSatisfatorio + (quantToleravel / 2)) / totalAmostras;
     }
 }
